requirements.txt
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-PyMongo==2.3.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
pymongo==3.9.0
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
